converted into a python package that can be pip installed and imported
