+++ requirements.txt
@@ -1,0 +1,3 @@
+colorzero==1.1
+gpiozero==1.5.1
+
