+++ setup.py
@@ -1,0 +1,10 @@
+from distutils.core import setup
+
+setup(
+    name='MinkaAireRemote',
+    version='0.1dev',
+    packages=['mar',],
+    license='MIT',
+    long_description=open('README.org').read(),
+)
+
